Create protocol for controller with uses TableAdapter
